own_MCP_servers/calculator/requirements.txt
# requirements.txt
fastmcp>=2.2.6
uvicorn[standard]>=0.22.0
fastapi>=0.95.0
python-dotenv>=1.0.0  # optional: helps local .env during development
